requirements.txt
Flask==2.3.3
Pillow==10.0.1
pytesseract==0.3.10
PyPDF2==3.0.1
pdf2image==1.16.3
python-dotenv==1.0.0